Proxy sendNative in sandboxed iframe back to main thread
Only the main thread can use native messaging, but we want to run most
of the code in a "sandbox" iframe, to allow for arbitrary eval for
https://github.com/deathcap/nodeachrome/issues/3 - so setup this
intricate messaging system to pass the results back and forth. Now this
works:

g.evalsb('ls()')

executing code in the sandbox, back to main, then to native. Data flow:

Main -> sandbox (eval)
Sandbox -> main -> native host