Rename 'sandbox' -> userland (OS kernel/userland analogy)
The term "sandbox" is way too overloaded. Chrome uses it to refer to an
iframe with a unique origin, that is therefore isolated with the rest of
the page, basicaly its own execution context, as detailed here:

https://developer.chrome.com/extensions/sandboxingEval

but "sandbox" is confusingly similar to "filesystem sandbox" yet
completely different, too generic. Eschew the sandbox verbiage in favor
of "userland processes", by analogy with the operating system kernel /
userland dichotomy. This should be much clearer conceptually.